- move the database tests to a subpackage
