+++ lending-database/src/test/scala/info/armado/ausleihe/database/access/LendIdentityCardDaoTest.scala
@@ -1,8 +1,8 @@
-package info.armado.ausleihe.database
+package info.armado.ausleihe.database.access
 
 import java.time.LocalDateTime
 
-import info.armado.ausleihe.database.access.{EnvelopeDao, IdentityCardDao, LendIdentityCardDao}
+import info.armado.ausleihe.database.WebDeployment
 import info.armado.ausleihe.database.barcode.Barcode
 import info.armado.ausleihe.database.entities._
 import javax.inject.Inject
